requirements.txt
httpx>=0.27.0
pydantic>=2.0
tiktoken>=0.7.0
pyyaml>=6.0
python-dotenv>=1.0
pymupdf>=1.24.0
click>=8.0
rich>=13.0
openai>=1.50.0
